Create conftest for common fixtures
