+++ tests/__init__.py
@@ -1,0 +1,10 @@
+import os
+
+from sqlmodel import create_engine
+
+url = (
+    f"postgresql://{os.environ['POSTGRES_USER']}:{os.environ['POSTGRES_PASSWORD']}"
+    f"@{os.environ['POSTGRES_HOST']}:{os.environ['POSTGRES_PORT']}/{os.environ['POSTGRES_DB']}"
+)
+
+engine = create_engine(url)
